removing unused imports and renaming components
