+++ webapp/src/App.tsx
@@ -1,17 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import {
-  CartItem,
-  NotificationType,
-  Order,
-  Product,
-  User,
-} from "./shared/shareddtypes";
 
 import NavBar from "./components/NavBar";
 import Footer from "./components/Footer";
-import Home from "./components/Home";
-import Shopping from "./components/Shopping";
+import Shop from "./components/Shop";
+import ShoppingCart from "./components/ShoppingCart";
 import SignIn from "./components/SignIn";
 import SignUp from "./components/SignUp";
 import Checkout from "./components/Checkout";
@@ -19,31 +12,30 @@ import ProductDetails from "./components/ProductDetails";
 import OrderDetails from "./components/OrderDetails";
 import OrderList from "./components/OrderList";
 
-import Snackbar from "@mui/material/Snackbar";
-import Alert from "@mui/material/Alert";
-
-import { getProducts } from "./api/api";
-import "bootstrap/dist/css/bootstrap.css";
-import { PayPalScriptProvider } from "@paypal/react-paypal-js";
-
-import "./App.css";
 import {
   createTheme,
   CssBaseline,
   PaletteMode,
   ThemeProvider,
   useMediaQuery,
+  Snackbar,
+  Alert,
 } from "@mui/material";
+import { grey, lightBlue } from "@mui/material/colors";
+
+import { getProducts } from "./api/api";
 import {
-  amber,
-  blue,
-  blueGrey,
-  deepOrange,
-  grey,
-  indigo,
-  lightBlue,
-  yellow,
-} from "@mui/material/colors";
+  CartItem,
+  NotificationType,
+  Product,
+  User,
+} from "./shared/shareddtypes";
+
+import { PayPalScriptProvider } from "@paypal/react-paypal-js";
+
+import "bootstrap/dist/css/bootstrap.css";
+
+import "./App.css";
 
 function App(): JSX.Element {
   const [notificationStatus, setNotificationStatus] = useState(false);
@@ -56,7 +48,6 @@ function App(): JSX.Element {
   const [productsCart, setProductsCart] = useState<CartItem[]>([]);
   const [totalUnitsInCart, setTotalUnitsInCart] = useState<number>(Number());
   const [user, setUser] = useState<User | null>(null);
-  const [orders, setOrders] = useState<Order[]>([]); //([{userId:'12', shippingPrice: 23,  totalPrice: 43}]);
 
   const createShop = async () => {
     const dbProducts: Product[] = await getProducts(); // and obtain the products
@@ -193,7 +184,7 @@ function App(): JSX.Element {
     localStorage.setItem("theme", String(initialTheme));
     console.log("none -> " + initialTheme);
   } else {
-    initialTheme = localStorage.getItem("theme") == "true";
+    initialTheme = localStorage.getItem("theme") === "true";
     console.log("already -> " + initialTheme);
   }
 
@@ -229,7 +220,7 @@ function App(): JSX.Element {
             <Route
               index
               element={
-                <Home
+                <Shop
                   products={products}
                   cartProducts={productsCart}
                   onAdd={handleAddCart}
@@ -239,7 +230,7 @@ function App(): JSX.Element {
             <Route
               path="cart"
               element={
-                <Shopping
+                <ShoppingCart
                   products={productsCart}
                   totalUnitsInCart={totalUnitsInCart}
                   onDecrementUnit={handleDecrementUnit}
@@ -271,13 +262,13 @@ function App(): JSX.Element {
                 <ProductDetails product={null as any} onAdd={handleAddCart} />
               }
             />
-          <Route
-            path="orders"
-            element={<OrderList userEmail={user?.email} />}
-          />
-          <Route path="/order/:code" element={<OrderDetails />} />
-        </Routes>
-        <Footer />
+            <Route
+              path="orders"
+              element={<OrderList userEmail={user?.email} />}
+            />
+            <Route path="/order/:code" element={<OrderDetails />} />
+          </Routes>
+          <Footer />
 
           <Snackbar
             open={notificationStatus}
