+++ webapp/src/components/ProductCommentList.tsx
@@ -1,14 +1,17 @@
-import React from "react";
-import {Grid, Paper} from "@mui/material";
-import {Review} from "../shared/shareddtypes";
+import { Grid, Paper } from "@mui/material";
+
+import { Review } from "../shared/shareddtypes";
+
 import ProductComment from "./ProductComment";
 
 type ProductListProps = {
-    reviews: Review[];
+  reviews: Review[];
 };
 
-function ProductCommentList(props: ProductListProps): JSX.Element {
-    /*const getCurrentCartAmount = (product: Product) => {
+export default function ProductCommentList(
+  props: ProductListProps
+): JSX.Element {
+  /*const getCurrentCartAmount = (product: Product) => {
         let currentAmount: number = 0;
         props.cartProducts.forEach((cartItem: CartItem) => {
           if (product.code === cartItem.product.code) {
@@ -18,19 +21,16 @@ function ProductCommentList(props: ProductListProps): JSX.Element {
         return currentAmount;
       };*/
 
-    return (
-            <Paper
-                style={{margin: "3vh 5vw", padding: "1em"}}>
-                <h2>User opinions about this product!</h2>
-                <Grid container columns={50} rowSpacing={5} className="mt-2 mb-2">
-                    {props.reviews.map((review: Review) => (
-                        <Grid item xs={25} key={(review.userEmail, review.productCode)}>
-                            <ProductComment review={review}/>
-                        </Grid>
-                    ))}
-                </Grid>
-            </Paper>
-    );
+  return (
+    <Paper style={{ margin: "3vh 5vw", padding: "1em" }}>
+      <h2>User opinions about this product!</h2>
+      <Grid container columns={50} rowSpacing={5} className="mt-2 mb-2">
+        {props.reviews.map((review: Review) => (
+          <Grid item xs={25} key={(review.userEmail, review.productCode)}>
+            <ProductComment review={review} />
+          </Grid>
+        ))}
+      </Grid>
+    </Paper>
+  );
 }
-
-export default ProductCommentList;
