+++ webapp/src/components/ProductList.tsx
@@ -1,42 +1,40 @@
-import {Card, Paper} from "@mui/material";
+import { Card } from "@mui/material";
 import Grid from "@mui/material/Grid";
 
-import {CartItem, Product} from "../shared/shareddtypes";
+import { CartItem, Product } from "../shared/shareddtypes";
+
 import ProductBox from "./ProductBox";
 
 type ProductListProps = {
-    products: Product[];
-    cartProducts: CartItem[];
-    OnAddCart: (product: Product) => void;
+  products: Product[];
+  cartProducts: CartItem[];
+  OnAddCart: (product: Product) => void;
 };
 
-function ProductList(props: ProductListProps): JSX.Element {
+export default function ProductList(props: ProductListProps): JSX.Element {
+  const getCurrentCartAmount = (product: Product) => {
+    let currentAmount: number = 0;
+    props.cartProducts.forEach((cartItem: CartItem) => {
+      if (product.code === cartItem.product.code) {
+        currentAmount = cartItem.amount;
+      }
+    });
+    return currentAmount;
+  };
 
-    const getCurrentCartAmount = (product: Product) => {
-        let currentAmount: number = 0;
-        props.cartProducts.forEach((cartItem: CartItem) => {
-            if (product.code === cartItem.product.code) {
-                currentAmount = cartItem.amount;
-            }
-        });
-        return currentAmount;
-    };
-
-    return (
-        <Grid container columns={50} rowSpacing={5} className="mt-2 mb-2">
-            {props.products.map((product: Product) => (
-                <Grid item xs={25} key={product.code}>
-                    <Card style={{margin: "1vh 8vw", padding: "1vh 1vw"}}>
-                        <ProductBox
-                            product={product}
-                            currentCartAmount={getCurrentCartAmount(product)}
-                            onAdd={(productToAdd: Product) => props.OnAddCart(productToAdd)}
-                        />
-                    </Card>
-                </Grid>
-            ))}
+  return (
+    <Grid container columns={50} rowSpacing={5} className="mt-2 mb-2">
+      {props.products.map((product: Product) => (
+        <Grid item xs={25} key={product.code}>
+          <Card style={{ margin: "1vh 8vw", padding: "1vh 1vw" }}>
+            <ProductBox
+              product={product}
+              currentCartAmount={getCurrentCartAmount(product)}
+              onAdd={(productToAdd: Product) => props.OnAddCart(productToAdd)}
+            />
+          </Card>
         </Grid>
-    );
+      ))}
+    </Grid>
+  );
 }
-
-export default ProductList;
